UI: update form handling
